refactor: 셀렉 박스 파일 구조 변경
