+++ src/components/common/Select/SelectTab/SelectTab.tsx
@@ -1,8 +1,8 @@
 import { RefObject, useEffect, useState } from 'react';
 import SelectTrigger from '../SelectTrigger/SelectTrigger';
 import { SelectTabContainer } from './SelectTab.style';
-import { useSelectOptionContext } from '../../../hooks/select/useSelectOptionContext';
-import { TRAVEL_LABEL } from '../../../constants';
+import { useSelectOptionContext } from '../../../../hooks/select/useSelectOptionContext';
+import { TRAVEL_LABEL } from '../../../../constants';
 
 interface SelectTabProps {
   isOpen: boolean;
